Add files via upload
The client's goal was to gain insights into the company's dataset. To enable me do a proper analysis, using excel, I had to answer some questions using the dataset provided by the company.
Converted the dataset to an excel table, created pivot tables and created an interactive dashboard. Below are a summary of some of the  insights:
-Majority of the customers are male, and most of the company's profit came from the young adult group.
-Company sells a wide range of products, phones recorded the product with highest cost price and revenue, but the least profit while accessories recorded the highest profit.
- Company made a profit of about 2.2million from July 2015 to July 2016.